moved good samaritan calls to driver
